updateClosedData
